Add overflow guards and fix allocation issues
Add comprehensive overflow guards for size_t arithmetic in array, bit_array,
string, pool, and arena allocators to prevent integer overflow
vulnerabilities. Enforce maximum 16-byte alignment in arena allocator.

Fix arena allocation to reuse existing chunks before creating new ones and
correctly align allocations based on absolute chunk base address. Fix array
and arena_array clone operations to preserve lifecycle hooks and copy
callbacks.

Improve chunk remapping to handle sequence markers correctly for legacy save
formats. Add proper object ID mapping for file_version < 7 in save pipeline.
Add nmo_error_free for releasing non-arena error chains.

Add extensive test coverage for overflow scenarios, deep copy verification,
lifecycle preservation, and type registry operations.

+++ src/app/parser.c
@@ -796,7 +796,7 @@ static int nmo_load_file_with_io(
         int32_t signed_raw_id = (int32_t) raw_id;
         if (id_sanitizer != NULL && signed_raw_id < 0) {
             int32_t runtime_ext = nmo_id_register_external(id_sanitizer, signed_raw_id);
-            if (runtime_ext != NMO_OBJECT_ID_INVALID) {
+            if (runtime_ext != (int32_t)NMO_OBJECT_ID_INVALID) {
                 nmo_log(logger, NMO_LOG_INFO,
                         "  Object %zu: external reference %d registered as runtime %d",
                         i, signed_raw_id, runtime_ext);
